+++ src/server/server_game.h
@@ -42,7 +42,7 @@ private:
   std::chrono::duration<float> rate = std::chrono::duration<float>((float)RATE);
 
 public:
-  explicit Game(int &game_id);
+  explicit Game(uint8_t &game_id);
 
   Queue<std::shared_ptr<RunnableCommandGame>> *
   add_player(std::shared_ptr<Queue<GameState>> sender_queue,
